Add a Navigation File Reader
The intent of this class is to parse the contents of an Antora nav.adoc
file. This file contains the navigation tree that Antora uses to
generate it's navigation links in the generated documentation. The
intent of this class isn't meant to be too detailed, just to be able to
read and parse the file and return any valid links.