Add support for a templating layer
This commit uses a combination of zend-expressive-twigrenderer,
zend-expressive-template, and twig to provide a template layer for the
library. At first blush, that might seem like a bit of overkill; and
perhaps it is. But the combination of these libraries makes the code
easier to write and maintain imho, as zend-expressive-twigrenderer
provides a unified interface to the template layer, making it trivial to
code against an interface, and not an implementation.