Add support for Symfony Console and zend-expressive-twigRenderer
The addition of these two packages will make it easier to use the
library from the console and to create file content using content
templates.

While the content could be created with homegrown scripts, it
makes no sense to do so, when tested packages already exist with
excellent documentation, such as Twig.

While the console/commandline scripts could be written by hand, it again
makes no sense to do so, when a well written and tested package already
exist, one that is effectively the PHP defacto standard.